+++ flowcraft/generator/components/annotation.py
@@ -76,7 +76,7 @@ class CardRgi(Process):
             - ``input_type``: fasta
             - ``output_type``: txt
             - ``ptype``: resistance gene detection (assembly)
-        """
+    """
 
     def __init__(self, **kwargs):
 
@@ -156,5 +156,51 @@ class Prokka(Process):
                 "cpus": 2,
                 "container": "ummidock/prokka",
                 "version": "1.12"
+            }
+        }
+
+
+class Diamond(Process):
+    """diamond process for protein database queries
+
+        This process is set with:
+
+            - ``input_type``: fasta
+            - ``output_type``: None
+            - ``ptype``: post_assembly
+    """
+
+    def __init__(self, **kwargs):
+
+        super().__init__(**kwargs)
+
+        self.input_type = "fasta"
+        self.output_type = None
+
+        self.params = {
+            "pathToDb": {
+                "default": 'null',
+                "description": "Provide full path for the diamond database. "
+                               "If none is provided then will try to fetch from"
+                               " the previous process. Default: None"
+            },
+            "fastaToDb": {
+                "default": 'null',
+                "description": "Provide the full path for the fasta to "
+                               "construct a diamond database. Default: None"
+            },
+            "blastType": {
+                "default": "'blastx'",
+                "description": "Defines the type of blast that diamond will do."
+                               "Can wither be blastx or blastp. Default: blastx"
+            }
+        }
+
+        self.directives = {
+            "diamond": {
+                "container": "flowcraft/diamond",
+                "version": "0.9.22-1",
+                "memory": "{ 4.GB * task.attempt }",
+                "cpus": 2
             }
         }